requirements.txt
aniso8601==7.0.0
ansicon==1.89.0
arrow==1.2.1
asgiref==3.4.1
awsebcli==3.20.3
bcrypt==3.2.0
beautifulsoup4==4.10.0
bitcoin==1.1.42
blessed==1.19.0
blis==0.7.5
botocore==1.23.49
catalogue==2.0.6
cement==2.8.2
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.3
cymem==2.0.6
Django==3.2.9
django-bootstrap-form==3.4
django-bootstrap3==21.2
django-cors-headers==3.10.1
django-crispy-forms==1.13.0
django-environ==0.8.1
django-excel==0.0.10
django-extensions==3.1.5
django-filter==21.1
django-livereload-server==0.3.2
django-mathfilters==1.0.0
django-picklefield==3.0.1
django-q==1.3.9
django-sendsms==0.5
django-tables2==2.4.1
django-widget-tweaks==1.4.12
djongo==1.3.6
dnspython==2.1.0
future==0.16.0
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
idna==3.3
Jinja2==3.0.3
jinxed==1.1.0
jmespath==0.10.0
joblib==1.1.0
langcodes==3.3.0
livereload==2.6.3
lml==0.1.0
mailer==0.8.1
MarkupSafe==2.0.1
mongoengine==0.23.1
murmurhash==1.0.6
numpy==1.21.5
packaging==21.3
pandas==1.3.5
pathspec==0.9.0
pathy==0.6.1
preshed==3.0.6
promise==2.3
psycopg2==2.9.3
pycparser==2.21
pydantic==1.8.2
pyexcel==0.6.7
pyexcel-io==0.6.5
pyexcel-webio==0.1.4
PyJWT==2.3.0
pymongo==3.12.3
pyparsing==3.0.6
pypiwin32==223
python-dateutil==2.8.2
pytz==2021.3
pywin32==303
PyYAML==5.4.1
redis==3.5.3
requests==2.26.0
Rx==1.6.1
scikit-learn==1.0.2
scipy==1.7.3
semantic-version==2.8.5
singledispatch==3.7.0
six==1.14.0
sklearn==0.0
smart-open==5.2.1
soupsieve==2.3.1
spacy==3.2.1
spacy-legacy==3.0.8
spacy-loggers==1.0.1
sqlparse==0.2.4
srsly==2.4.2
stripe==2.65.0
termcolor==1.1.0
text-unidecode==1.3
texttable==1.6.4
thinc==8.0.13
threadpoolctl==3.0.0
tornado==6.1
tqdm==4.62.3
twilio==7.5.0
typer==0.4.0
typing_extensions==4.0.1
urllib3==1.26.8
wasabi==0.9.0
wcwidth==0.1.9
